TFS 10971 - Survey
git-tfs-id: [https://tfs.omnicloud.local/tfs/cco%20program/]$/eCoaching_V2;C40011

+++ Code/Web Development/eCoachingLog/eCoachingLog/Models/Survey/Survey.cs
@@ -7,8 +7,8 @@ namespace eCoachingLog.Models.Survey
 {
 	public class Survey
 	{
-		public int Id { get; set; }
-		public int LogId { get; set; }
+		public long Id { get; set; }
+		public long LogId { get; set; }
 		public string LogName { get; set; }
 		public string EmployeeId { get; set; }
 		public IList<Question> Questions { get; set; }
